Refactorization of Signup Screens (#83)
* Changed the file structure for signup screens

* Import necessary resources

* Create a reusable component for signup text

* Invoke SignUpText

* Renamed to SharedComposables

* Moved to SharedComposables.kt

* Import necessary resources

* Add SignUpText

* Add SignUpText

* Change to a more descriptive name

* Add comment

* Import resources

* Add reusable placeholder composable

* Add reusable placeholder composable

* Replace with the reusable one

* Make padding concise

* Import resources

* Refactor SignupText

* Refactor SignupText

* Refactor unnecesarry Column

* Add keyboardOption for email input

* Remove import

* Remove unnecessary Column

* Refactor InputField
