+++ app/src/main/java/com/it235/nureserved/composables/SharedComposables.kt
@@ -20,14 +20,19 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.it235.nureserved.R
 import com.it235.nureserved.ScreenRoutes
+import com.it235.nureserved.font.poppinsFamily
 import com.it235.nureserved.ui.theme.brandColorBlue
+import com.it235.nureserved.ui.theme.darkGray
+import com.it235.nureserved.ui.theme.white3
 
 @Composable
 fun RowHeader(textHeader: String = "ROOM RESERVATIONS", modifier: Modifier = Modifier.width(10.dp)){
@@ -92,6 +97,43 @@ private fun CustomDialog(
         }
     )
 }
+
+/* SignUp Screens Composable */
+
+@Composable
+fun SignUpText(
+    modifier: Modifier = Modifier,
+    text: String,
+    fontSize: TextUnit,
+    textAlign: TextAlign? = null,
+){
+    Text(
+        modifier = modifier,
+        color = darkGray,
+        text = text,
+        textAlign = textAlign,
+        style = TextStyle(
+            fontFamily = poppinsFamily,
+            fontWeight = FontWeight.SemiBold,
+            fontSize = fontSize,
+        )
+    )
+}
+
+@Composable
+fun AuthInputPlaceholderTextStyle(label: String){
+    Text(
+        color = white3,
+        text = label,
+        style = TextStyle(
+            fontFamily = poppinsFamily,
+            fontSize = 18.sp,
+            fontWeight = FontWeight.Medium,
+        )
+    )
+}
+
+/* end of SignUp Screens Composable */
 
 @Composable
 fun SuccessDialog(
